refactor V2 root
