+++ packages/portal-proto/src/features/files/FilesView.tsx
@@ -132,7 +132,7 @@ export const ContextualFilesView: React.FC = () => {
           <DownloadIcon size={"1.5rem"} />
           Manifest
         </Button>
-        <Link href="/user-flow/workbench/MultipleImageViewerPage" passHref>
+        <Link href="/image-viewer/MultipleImageViewerPage" passHref>
           <Button component="a" className={buttonStyle}>
             View Images
           </Button>
