+++ packages/portal-proto/src/features/repositoryApp/RepositoryApp.tsx
@@ -136,7 +136,7 @@ const RepositoryApp = () => {
             <DownloadIcon size={"1.5rem"} />
             Manifest
           </FunctionButton>
-          <Link href="/user-flow/workbench/MultipleImageViewerPage">
+          <Link href="/image-viewer/MultipleImageViewerPage">
             <FunctionButton component="a">View Images</FunctionButton>
           </Link>
         </div>
