+++ packages/portal-proto/src/pages/image-viewer/MultipleImageViewerPage.tsx
@@ -1,6 +1,6 @@
 import { NextPage } from "next";
-import { MultipleImageViewer } from "../../../components/MultipleImageViewer";
-import { UserFlowVariedPages } from "../../../features/layout/UserFlowVariedPages";
+import { MultipleImageViewer } from "@/components/MultipleImageViewer";
+import { UserFlowVariedPages } from "@/features/layout/UserFlowVariedPages";
 import { headerElements } from "@/features/user-flow/workflow/navigation-utils";
 import { useRouter } from "next/router";
 
@@ -9,7 +9,7 @@ const MultipleImageViewerPage: NextPage = () => {
   return (
     <UserFlowVariedPages
       {...{
-        indexPath: "/user-flow/workbench/analysis_page?app=",
+        indexPath: "/analysis_page?app=",
         headerElements,
       }}
     >
